systemd: expand template before start
Replace files updated by upstream package.

ReadWriteDirectories is used to make /etc/netdata directory
writable by the ExecStatPre commands.
Without this option, /etc/netdata directory would be read-only
for the systemd unit.
This is not a security issue, because the process is started using netdata
user and it will not be able to overwrite any file inside the dir.